+++ tree.go
@@ -1,0 +1,278 @@
+package merkletree
+
+import (
+	"errors"
+	"hash"
+)
+
+// A Tree takes data as leaves and returns the Merkle root. Each call to 'Push'
+// adds one leaf to the Merkle tree. Calling 'Root' returns the Merkle root.
+// The Tree also constructs proof that a single leaf is a part of the tree. The
+// leaf can be chosen with 'SetIndex'. The memory footprint of Tree grows in
+// O(log(n)) in the number of leaves.
+type Tree struct {
+	// The Tree is stored as a stack of subtrees. Each subtree has a height,
+	// and is the Merkle root of 2^height leaves. A Tree with 11 nodes is
+	// represented as a subtree of height 3 (8 nodes), a subtree of height 1 (2
+	// nodes), and a subtree of height 0 (1 node). Head points to the smallest
+	// tree. When a new leaf is inserted, it is inserted as a subtree of height
+	// 0. If there is another subtree of the same height, both can be removed,
+	// combined, and then inserted as a subtree of height n + 1.
+	head *subTree
+	hash hash.Hash
+
+	// Helper variables used to construct proofs that the data at 'proofIndex'
+	// is in the Merkle tree. The proofSet is constructed as elements are being
+	// added to the tree. The first element of the proof set is the original
+	// data used to create the leaf at index 'proofIndex'.
+	currentIndex uint64
+	proofIndex   uint64
+	proofSet     [][]byte
+
+	// The cachedTree flag indicates that the tree is cached, meaning that
+	// different code is used in 'Push' for creating a new head subtree. Adding
+	// this flag is somewhat gross, but eliminates needing to duplicate the
+	// entire 'Push' function when writing the cached tree.
+	cachedTree bool
+}
+
+// A subTree contains the Merkle root of a complete (2^height leaves) subTree
+// of the Tree. 'sum' is the Merkle root of the subTree. If 'next' is not nil,
+// it will be a tree with a higher height.
+type subTree struct {
+	next   *subTree
+	height int
+	sum    []byte
+}
+
+// sum returns the hash of the input data using the specified algorithm.
+func sum(h hash.Hash, data []byte) []byte {
+	if data == nil {
+		return nil
+	}
+
+	_, err := h.Write(data)
+	if err != nil {
+		// Result will not be correct if the hash is throwing an error when
+		// it's supposed to be checksumming data.
+		panic(err)
+	}
+	result := h.Sum(nil)
+	h.Reset()
+	return result
+}
+
+// leafSum returns the hash created from data inserted to form a leaf. Leaf
+// sums are calculated using:
+//		Hash( 0x00 || data)
+func leafSum(h hash.Hash, data []byte) []byte {
+	return sum(h, append([]byte{0}, data...))
+}
+
+// nodeSum returns the hash created from two sibling nodes being combined into
+// a parent node. Node sums are calculated using:
+//		Hash( 0x01 || left sibling sum || right sibling sum)
+func nodeSum(h hash.Hash, a, b []byte) []byte {
+	return sum(h, append(append([]byte{1}, a...), b...))
+}
+
+// joinSubTrees combines two equal sized subTrees into a larger subTree.
+func joinSubTrees(h hash.Hash, a, b *subTree) *subTree {
+	if DEBUG {
+		if b.next != a {
+			panic("invalid subtree join - 'a' is not paired with 'b'")
+		}
+		if a.height < b.height {
+			panic("invalid subtree presented - height mismatch")
+		}
+	}
+
+	return &subTree{
+		next:   a.next,
+		height: a.height + 1,
+		sum:    nodeSum(h, a.sum, b.sum),
+	}
+}
+
+// New initializes a Tree with a hash object, which will be used when hashing
+// the input.
+func New(h hash.Hash) *Tree {
+	return &Tree{
+		hash: h,
+	}
+}
+
+// Push adds a leaf to the tree by hashing the input and then inserting the
+// result as a leaf.
+func (t *Tree) Push(data []byte) {
+	// The first element of a proof is the data at the proof index. If this
+	// data is being inserted at the proof index, it is added to the proof set.
+	if t.currentIndex == t.proofIndex {
+		t.proofSet = append(t.proofSet, data)
+	}
+
+	// Hash the data to create a subtree of height 0. The sum of the new node
+	// is going to be the data for cached trees, and is going to be the result
+	// of calling leafSum() on the data for standard trees. Doing a check here
+	// prevents needing to duplicate the entire 'Push' function for the trees.
+	t.head = &subTree{
+		next:   t.head,
+		height: 0,
+	}
+	if t.cachedTree {
+		t.head.sum = data
+	} else {
+		t.head.sum = leafSum(t.hash, data)
+	}
+
+	// Insert the subTree into the Tree. As long as the height of the next
+	// subTree is the same as the height of the current subTree, the two will
+	// be combined into a single subTree of height n+1.
+	for t.head.next != nil && t.head.height == t.head.next.height {
+		// Before combining subtrees, check whether one of the subtree hashes
+		// needs to be added to the proof set. This is going to be true IFF the
+		// subtrees being combined are one height higher than the previous
+		// subtree added to the proof set. The height of the previous subtree
+		// added to the proof set is equal to len(t.proofSet) - 1.
+		if t.head.height == len(t.proofSet)-1 {
+			// One of the subtrees needs to be added to the proof set. The
+			// subtree that needs to be added is the subtree that does not
+			// contain the proofIndex. Because the subtrees being compared are
+			// the smallest and rightmost trees in the Tree, this can be
+			// determined by rounding the currentIndex down to the number of
+			// nodes in the subtree and comparing that index to the proofIndex.
+			leaves := uint64(1 << uint(t.head.height))
+			mid := (t.currentIndex / leaves) * leaves
+			if t.proofIndex < mid {
+				t.proofSet = append(t.proofSet, t.head.sum)
+			} else {
+				t.proofSet = append(t.proofSet, t.head.next.sum)
+			}
+
+			// Sanity check - the proofIndex should never be less than the
+			// midpoint minus the number of leaves in each subtree.
+			if DEBUG {
+				if t.proofIndex < mid-leaves {
+					panic("proof being added with weird values")
+				}
+			}
+		}
+
+		// Join the two subTrees into one subTree with a greater height. Then
+		// compare the new subTree to the next subTree.
+		t.head = joinSubTrees(t.hash, t.head.next, t.head)
+	}
+	t.currentIndex++
+
+	// Sanity check - From head to tail of the stack, the height should be
+	// strictly increasing.
+	if DEBUG {
+		current := t.head
+		height := current.height
+		for current.next != nil {
+			current = current.next
+			if current.height <= height {
+				panic("subtrees are out of order")
+			}
+			height = current.height
+		}
+	}
+}
+
+// Root returns the Merkle root of the data that has been pushed into the Tree.
+func (t *Tree) Root() []byte {
+	// If the Tree is empty, return the hash of the empty string.
+	if t.head == nil {
+		return sum(t.hash, nil)
+	}
+
+	// The root is formed by hashing together subTrees in order from least in
+	// height to greatest in height. The taller subtree is the first subtree in
+	// the join.
+	current := t.head
+	for current.next != nil {
+		current = joinSubTrees(t.hash, current.next, current)
+	}
+	return current.sum
+}
+
+// Prove returns a proof that the data at index 'proofIndex' is an element in
+// the current Tree. The proof will be invalid if any more elements are added
+// to the tree after calling Prove. The tree is left unaltered.
+func (t *Tree) Prove() (merkleRoot []byte, proofSet [][]byte, proofIndex uint64, numLeaves uint64) {
+	// Return nil if the Tree is empty, or if the proofIndex hasn't yet been
+	// reached.
+	if t.head == nil || len(t.proofSet) == 0 {
+		return t.Root(), nil, t.proofIndex, t.currentIndex
+	}
+	proofSet = t.proofSet
+
+	// The set of subtrees must now be collapsed into a single root. The proof
+	// set already contains all of the elements that are members of a complete
+	// subtree. Of what remains, there will be at most 1 element provided from
+	// a sibling on the right, and all of the other proofs will be provided
+	// from a sibling on the left. This results from the way orphans are
+	// treated. All subtrees smaller than the subtree containing the proofIndex
+	// will be combined into a single subtree that gets combined with the
+	// proofIndex subtree as a single right sibling. All subtrees larger than
+	// the subtree containing the proofIndex will be combined with the subtree
+	// containing the proof index as left siblings.
+
+	// Start at the smallest subtree and combine it with larger subtrees until
+	// it would be combining with the subtree that contains the proof index. We
+	// can recognize the subtree containing the proof index because the height
+	// of that subtree will be one less than the current length of the proof
+	// set.
+	current := t.head
+	for current.next != nil && current.next.height < len(proofSet)-1 {
+		current = joinSubTrees(t.hash, current.next, current)
+	}
+
+	// Sanity check - check that either 'current' or 'current.next' is the
+	// subtree containing the proof index.
+	if DEBUG {
+		if current.height != len(t.proofSet)-1 && (current.next != nil && current.next.height != len(t.proofSet)-1) {
+			panic("could not find the subtree containing the proof index")
+		}
+	}
+
+	// If the current subtree is not the subtree containing the proof index,
+	// then it must be an aggregate subtree that is to the right of the subtree
+	// containing the proof index, and the next subtree is the subtree
+	// containing the proof index.
+	if current.next != nil && current.next.height == len(proofSet)-1 {
+		proofSet = append(proofSet, current.sum)
+		current = current.next
+	}
+
+	// The current subtree must be the subtree containing the proof index. This
+	// subtree does not need an entry, as the entry was created during the
+	// construction of the Tree. Instead, skip to the next subtree.
+	current = current.next
+
+	// All remaning subtrees will be added to the proof set as a left sibling,
+	// completeing the proof set.
+	for current != nil {
+		proofSet = append(proofSet, current.sum)
+		current = current.next
+	}
+	return t.Root(), proofSet, t.proofIndex, t.currentIndex
+}
+
+// Reset returns the tree to its inital, empty state.
+func (t *Tree) Reset() {
+	t.head = nil
+	t.currentIndex = 0
+	t.proofIndex = 0
+	t.proofSet = nil
+}
+
+// SetIndex must be called on an empty Tree. Trees can be emptied by calling
+// Reset.
+func (t *Tree) SetIndex(i uint64) error {
+	if t.head != nil {
+		return errors.New("cannot call SetIndex on Tree if Tree has not been reset")
+	}
+	t.proofIndex = i
+	return nil
+}
